[refactor] Split scheduler into shared/ subfolder and .cc files
Move reusable scheduler base types (SchedulerOperationBase,
SchedulerSender, SchedulerHandle, StoppableSchedulerCompletionSignatures)
into scheduler/shared/ for clearer layering.

Extract non-template method bodies (constructors, EnqueueOperation,
WorkerThreadLoop, StartWorkers) from CorePinnedThreadPool,
PriorityThreadPool, and WorkSharingThreadPool into corresponding .cc
files to reduce header bloat.

Also fix a latent missing include of platform.h in network.cc that was
previously satisfied transitively through the pool headers.
