+++ include/ex_actor/internal/scheduler/shared/stoppable_scheduler_completion_signatures.h
